refactor(chat-ui): extract ChatIntegrationPill, dedupe the two call sites
The Status card and the agent-home tag both rendered their own status pill off
the shared label/color constants. Extract a single ChatIntegrationPill that owns
the label + dot treatment, keyed on a derived ChatIntegrationState, and use it in
both places. Visuals unchanged: card = dot/pulse + text-xs, home tag = plain
text-2xs (size + showDot props).
